political_news/requirements.txt
Flask[async]
openai
requests
pyyaml
Flask-CORS
aiohttp